+++ evalml/tests/automl_tests/test_automl.py
@@ -2262,15 +2262,12 @@ def test_time_series_regression_with_parameters(ts_data):
     assert automl.allowed_pipelines[0].parameters["pipeline"] == problem_configuration
 
 
-@pytest.mark.parametrize("graph_type", ["linear", "dict", "cg"])
+@pytest.mark.parametrize("graph_type", ["dict", "cg"])
 def test_automl_accepts_component_graphs(
     graph_type, dummy_classifier_estimator_class, X_y_binary
 ):
     X, y = X_y_binary
-    if graph_type == "linear":
-        component_graph = ["Imputer", "Logistic Regression Classifier"]
-        component_graph_obj = ComponentGraph.from_list(component_graph)
-    elif graph_type == "dict":
+    if graph_type == "dict":
         component_graph = {
             "imputer": ["Imputer"],
             "ohe": ["One Hot Encoder", "imputer.x"],
