+++ evalml/tests/pipeline_tests/test_pipelines.py
@@ -7,6 +7,7 @@ import pandas as pd
 import pytest
 import woodwork as ww
 from pandas.testing import assert_frame_equal
+from skopt.space import Categorical, Integer, Real
 
 from evalml.exceptions import (
     MissingComponentError,
@@ -2303,3 +2304,32 @@ def test_oversampler_component_in_pipeline_predict(oversampler):
     assert len(preds) == 1000
     preds = pipeline.predict_proba(X)
     assert len(preds) == 1000
+
+
+def test_get_hyperparameter_ranges():
+    pipeline = BinaryClassificationPipeline(
+        component_graph=["Imputer", "Random Forest Classifier"]
+    )
+    custom_hyperparameters = {
+        "One Hot Encoder": {"top_n": 3},
+        "Imputer": {"numeric_impute_strategy": Categorical(["most_frequent", "mean"])},
+        "Random Forest Classifier": {"n_estimators": Integer(150, 160)},
+    }
+
+    algo_ranges = {
+        "Imputer": {
+            "categorical_impute_strategy": ["most_frequent"],
+            "numeric_impute_strategy": Categorical(
+                categories=("most_frequent", "mean"), prior=None
+            ),
+        },
+        "Random Forest Classifier": {
+            "n_estimators": Integer(
+                low=150, high=160, prior="uniform", transform="identity"
+            ),
+            "max_depth": Integer(low=1, high=10, prior="uniform", transform="identity"),
+        },
+    }
+    hyper_ranges = pipeline.get_hyperparameter_ranges(custom_hyperparameters)
+
+    assert algo_ranges == hyper_ranges
